pyproject.toml
[project]
name = "spaceweather"
version = "0.1.1"
description = ""
authors = [
    {name = "DJ Stomp",email = "85457381+DJStompZone@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx (>=0.28.1,<0.29.0)",
    "tqdm (>=4.67.1,<5.0.0)",
    "pillow (>=11.3.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
